+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1 (unreleased)
 
 - Improved typecasting
+- Removed table prefix with Hive
 
 ## 0.1.0 (2021-01-24)